clamp dates with invalid years to 0
Asp er the MySQL documentation DATE and DATETIME only support ranges from
'1000-01-01 00:00:00' to '9999-12-31 23:59:59', and TIMESTAMP from
'1970-01-01 00:00:01.000000' to '2038-01-19 03:14:07.999999'. Values outside
those ranges are not guaranteed to work. Additionally it mentions that invalid
values are converted to their appropriate zero value.

The patch converts dates that lead to an exception right now to their zero
value. This is more lenient than what the ranges given in the documentation.
The behaviour set through zeroDateTimeBehavior applies to these zero values.

This fixes Bug#94872.